Fix crash of laser_assembler if service called before receiving scans
Although this is extremely unlikely to hit on an actual robot, this code
would SEGFAULT if the service was called before any scans had been
received.

The code that crashes is a few lines down:
`scan_hist_.rbegin()->header.stamp`.

I ran into this when trying to run the tilt lidar nodes against a bag file.
In that case it is very easy to hit this problem.

This bug does not exist in newer branches of the repo since this code has
been re-written, therefore this change does not need to be upstreamed.