Fix crash of laser_assembler if service called before receiving scans
Although this is extremely unlikely to hit on an actual robot, this code
would SEGFAULT if the service was called before any scans had been
received.

I ran into this when trying to run the tilt lidar nodes against a bag file.
In that case it is very easy to hit this problem.

This does not need to be upstreamed since the bug is in code that has not
been upstreamed.