Fix crash of laser_assembler if service called before receiving scans
Although this is extremely unlikely to hit on an actual robot, this code
would SEGFAULT if the service was called before any scans had been
received.

I ran into this when trying to run the tilt lidar nodes against a bag file.
In that case it is very easy to hit this problem.

This commit should be squashed with the previous 'add service
assemble_scans2_blocking' if it is ever upstreamed.